+++ WalletWasabi.Fluent/Services.cs
@@ -3,6 +3,7 @@ using WalletWasabi.Helpers;
 using WalletWasabi.Services;
 using WalletWasabi.Stores;
 using WalletWasabi.Tor;
+using WalletWasabi.Tor.StatusChecker;
 using WalletWasabi.Wallets;
 using WalletWasabi.WebClients.Wasabi;
