+++ WalletWasabi.Fluent/ViewModels/StatusIcon/IStatusIconViewModel.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 using System.Windows.Input;
-using WalletWasabi.Services.Tor.StatusChecker;
+using WalletWasabi.Tor.StatusChecker;
 
 namespace WalletWasabi.Fluent.ViewModels.StatusIcon;
