+++ WalletWasabi.Fluent/ViewModels/StatusIcon/StatusIconDesignViewModel.cs
@@ -1,7 +1,7 @@
 using ReactiveUI;
 using System.Collections.Generic;
 using System.Windows.Input;
-using WalletWasabi.Services.Tor.StatusChecker;
+using WalletWasabi.Tor.StatusChecker;
 
 namespace WalletWasabi.Fluent.ViewModels.StatusIcon;
