+++ WalletWasabi/Tor/StatusChecker/XmlIssueListParser.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Xml;
 
-namespace WalletWasabi.Services.Tor.StatusChecker;
+namespace WalletWasabi.Tor.StatusChecker;
 
 public class XmlIssueListParser
 {
