+++ WalletWasabi/WalletWasabi.csproj
@@ -80,4 +80,8 @@
 		<RuntimeHostConfigurationOption Include="System.Globalization.Invariant" Value="true" />
 	</ItemGroup>
 
+	<ItemGroup>
+	  <Folder Include="Services\Tor\" />
+	</ItemGroup>
+
 </Project>
